requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
boto3>=1.34.0
botocore>=1.34.0
pydantic>=2.6.0
PyYAML>=6.0.1
Pillow>=10.0.0
numpy>=1.26.0
tenacity>=8.2.3

